+++ src/app/util/interfaces/damageTypeMap.tsx
@@ -1,0 +1,13 @@
+export default interface DamageTypeMap<T> {
+    [key: string]: T | undefined;
+    physical: T;
+    magic: T;
+    fire: T;
+    lightning: T;
+    holy: T;
+    poison?: T;
+    bleed?: T;
+    sleep?: T;
+    madness?: T;
+    "scarlet-rot"?: T;
+}