+++ src/app/util/types/correction.tsx
@@ -1,6 +1,6 @@
 export type Correction = {
     id: string;
-    softcaps: number[][];
-    growth: number[][];
-    adjustments: number[][];
+    softcaps: number[];
+    growth: number[];
+    adjustments: number[];
 };